git: add llm-pick
When backporting patches, cherry-picking sometimes fails due to slight
differences in the diff context. The patch tool allows to be more fuzzy.
To be able to easily backport more commits fully automated, this commit
introduces a git-llm--pick tool, which runs in order, on failure:

    1. git-cherry-pick first,
    2. tries pick after picking context commits
    3. uses the patch tool
    4. uses an LLM model to adjust the code.

The tool supports the basic operations of git-cherry-pick, and warns in
case parameters are used that the tool does not know.

To improve confidence in the change, the tool can run a validation
command after the modification. The change is rolled back in case the
validation fails. This way, confidence in the change is increased.
Additionally, limits for the LLM output can be specified.

When using the tool to perform backporting, the amount of patches
that can be picked automatically increases. We test backporting patches
from  Linux kernel 6.12 to kernel 6.1. Specifically, when re-backporting
commits with "Fixes:" lines from 6.1.142 to 6.1.147, and reapplying to
the release 6.1.142, git-llm-pick increases the number of patches to
picl automatically by 17% (7% backport, 10% LLM). The below shows the
distribution of patches that can be picked with the git-fuzzy-pick tool:

Patches attempted to backport: 261
Patches backported successfully: 187
Backported total: 200
Applied with patch tool:
     14     Applied with LLM-adjusted hunks for 1 functions from us.amazon.nova
      1     Applied with LLM-adjusted hunks for 2 functions from us.amazon.nova
      1     Applied with LLM-adjusted hunks for 3 functions from us.amazon.nova
Picked with help of context commits: 12

All modifications have been validated by attempting to build the object
files of the modified C source files with `make` with the allyes
config. For header files, we compiled the first source file that
included the file.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>