llm,scripts: focus more on semantics
When analyzing the backporting results, we noticed that the LLM
sometimes misses lines that should be removed. This commit applies
prompt engineering to the prompt to highlight the difference in
those lines. The change is applied for the instructions for the
two tasks of adapting hunks and adapting sections.

To pass testing, the cached LLM responses have been updated accordingly.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>