git,commit: choose explanation
To be able to reduce the human interaction per picked commits, this
change allows to change the amount of information to be put into the
commit message. This way, depending on project we can set a default.
Afterwards, a human user can still rewrite the commit message to meet
the projects expectations.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>